[FX] Move Delegate to its own file
[ghstack-poisoned]
