+++ torch/fx/__init__.py
@@ -85,8 +85,9 @@ Because this code is valid PyTorch code, the resulting `GraphModule` can be used
 `nn.Module` can be used, including in TorchScript tracing/compilation.
 '''
 
+from .delegate import DefaultDelegate
 from .graph_module import GraphModule
-from .symbolic_trace import symbolic_trace, DefaultDelegate, ModuleHierarchyCtxMgr
+from .symbolic_trace import symbolic_trace, ModuleHierarchyCtxMgr
 from .graph import Graph, map_arg
 from .node import Node
 from .proxy import Proxy
